# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type GmailAnalysisInputMessage = {
  /**
     * @minLength 1
     * @maxLength 256
     */
  id: string;
  /** @maxLength 1000 */
  sender: string;
  /** @maxLength 2000 */
  subject: string;
  /** @maxLength 100 */
  date: string;
  /** @maxLength 10000 */
  snippet: string;
};

export interface GmailAnalysisInput {
  message: GmailAnalysisInputMessage;
}

/**
 * @nullable
 */
export type GmailExtractionOpportunityType = typeof GmailExtractionOpportunityType[keyof typeof GmailExtractionOpportunityType] | null;


export const GmailExtractionOpportunityType = {
  'Full-time': 'Full-time',
  'Part-time': 'Part-time',
  Internship: 'Internship',
  Training: 'Training',
  'Train-to-Hire': 'Train-to-Hire',
  Freelance: 'Freelance',
} as const;

/**
 * @nullable
 */
export type GmailExtractionStatus = typeof GmailExtractionStatus[keyof typeof GmailExtractionStatus] | null;


export const GmailExtractionStatus = {
  Saved: 'Saved',
  Applied: 'Applied',
  Under_Review: 'Under Review',
  Assessment: 'Assessment',
  Interview: 'Interview',
  Technical_Interview: 'Technical Interview',
  Final_Interview: 'Final Interview',
  Offer: 'Offer',
  Accepted: 'Accepted',
  Rejected: 'Rejected',
  Withdrawn: 'Withdrawn',
} as const;

export interface GmailExtraction {
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  position: string | null;
  /** @nullable */
  opportunityType: GmailExtractionOpportunityType;
  /** @nullable */
  status: GmailExtractionStatus;
  /** @nullable */
  applicationDate: string | null;
  /** @nullable */
  interviewDate: string | null;
  /** @nullable */
  interviewType: string | null;
  /** @nullable */
  interviewLocation: string | null;
  /** @nullable */
  interviewNotes: string | null;
  /** @nullable */
  jobUrl: string | null;
}

export interface GmailAnalysis {
  relevant: boolean;
  reason: string;
  /** @nullable */
  reviewToken: string | null;
  extraction: GmailExtraction;
}

export interface GmailSearchInput {
  /** @maxLength 500 */
  query?: string;
}

export interface ApiError {
  error: string;
}

export type ApplicationInputOpportunityType = typeof ApplicationInputOpportunityType[keyof typeof ApplicationInputOpportunityType];


export const ApplicationInputOpportunityType = {
  'Full-time': 'Full-time',
  'Part-time': 'Part-time',
  Internship: 'Internship',
  Training: 'Training',
  'Train-to-Hire': 'Train-to-Hire',
  Freelance: 'Freelance',
} as const;

export type ApplicationInputWorkMode = typeof ApplicationInputWorkMode[keyof typeof ApplicationInputWorkMode];


export const ApplicationInputWorkMode = {
  'On-site': 'On-site',
  Remote: 'Remote',
  Hybrid: 'Hybrid',
} as const;

export type ApplicationInputStatus = typeof ApplicationInputStatus[keyof typeof ApplicationInputStatus];


export const ApplicationInputStatus = {
  Saved: 'Saved',
  Applied: 'Applied',
  Under_Review: 'Under Review',
  Assessment: 'Assessment',
  Interview: 'Interview',
  Technical_Interview: 'Technical Interview',
  Final_Interview: 'Final Interview',
  Offer: 'Offer',
  Accepted: 'Accepted',
  Rejected: 'Rejected',
  Withdrawn: 'Withdrawn',
} as const;

export type ApplicationInputSource = typeof ApplicationInputSource[keyof typeof ApplicationInputSource];


export const ApplicationInputSource = {
  Manual: 'Manual',
  Gmail: 'Gmail',
} as const;

export interface ApplicationInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  companyName: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  positionTitle: string;
  opportunityType: ApplicationInputOpportunityType;
  workMode: ApplicationInputWorkMode;
  /**
     * Local calendar date (YYYY-MM-DD)
     * @pattern ^\d{4}-\d{2}-\d{2}$
     */
  applicationDate: string;
  status: ApplicationInputStatus;
  /** Empty string or YYYY-MM-DD */
  interviewDate: string;
  /** Empty string or HH:MM */
  interviewTime: string;
  /** @maxLength 255 */
  interviewType: string;
  /** @maxLength 2048 */
  interviewLocation: string;
  interviewNotes: string;
  /**
     * Empty string or an http(s) URL
     * @maxLength 2048
     */
  jobUrl: string;
  notes: string;
  source: ApplicationInputSource;
}

export type ApplicationOpportunityType = typeof ApplicationOpportunityType[keyof typeof ApplicationOpportunityType];


export const ApplicationOpportunityType = {
  'Full-time': 'Full-time',
  'Part-time': 'Part-time',
  Internship: 'Internship',
  Training: 'Training',
  'Train-to-Hire': 'Train-to-Hire',
  Freelance: 'Freelance',
} as const;

export type ApplicationWorkMode = typeof ApplicationWorkMode[keyof typeof ApplicationWorkMode];


export const ApplicationWorkMode = {
  'On-site': 'On-site',
  Remote: 'Remote',
  Hybrid: 'Hybrid',
} as const;

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  Saved: 'Saved',
  Applied: 'Applied',
  Under_Review: 'Under Review',
  Assessment: 'Assessment',
  Interview: 'Interview',
  Technical_Interview: 'Technical Interview',
  Final_Interview: 'Final Interview',
  Offer: 'Offer',
  Accepted: 'Accepted',
  Rejected: 'Rejected',
  Withdrawn: 'Withdrawn',
} as const;

export type ApplicationSource = typeof ApplicationSource[keyof typeof ApplicationSource];


export const ApplicationSource = {
  Manual: 'Manual',
  Gmail: 'Gmail',
} as const;

export interface Application {
  id: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  companyName: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  positionTitle: string;
  opportunityType: ApplicationOpportunityType;
  workMode: ApplicationWorkMode;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  applicationDate: string;
  status: ApplicationStatus;
  interviewDate: string;
  interviewTime: string;
  /** @maxLength 255 */
  interviewType: string;
  /** @maxLength 2048 */
  interviewLocation: string;
  interviewNotes: string;
  /** @maxLength 2048 */
  jobUrl: string;
  notes: string;
  source: ApplicationSource;
  /** Server-owned ISO 8601 timestamp */
  createdAt: string;
  /** Server-owned ISO 8601 timestamp */
  updatedAt: string;
}

export interface HealthStatus {
  status: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ApiError;

/**
 * Application not found
 */
export type NotFoundResponse = ApiError;

/**
 * Unexpected server error
 */
export type ServerErrorResponse = ApiError;

export type ConnectGmail200 = {
  connected: boolean;
};

export type SearchGmail200MessagesItem = {
  id: string;
  sender: string;
  subject: string;
  date: string;
  snippet: string;
};

export type SearchGmail200 = {
  hasMore: boolean;
  messages: SearchGmail200MessagesItem[];
};