# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: applications
    description: Shared JobTrack application records (authentication is not enabled)
paths:
  /gmail/analyze:
    post:
      operationId: analyzeGmail
      summary: Analyze one email preview without saving it (development only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GmailAnalysisInput"
      responses:
        "200":
          description: Classification and nullable extracted fields for user review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailAnalysis"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Not available outside development
        "429":
          description: Analysis busy or provider rate limited
        "502":
          description: AI returned no valid analysis
        "503":
          description: AI service unavailable
  /gmail/connect:
    post:
      operationId: connectGmail
      summary: Verify the Replit Gmail connection (development only)
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                type: object
                required: [connected]
                properties:
                  connected: { type: boolean }
  /gmail/search:
    post:
      operationId: searchGmail
      summary: Manually retrieve up to 20 Gmail previews (development only)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GmailSearchInput"
      responses:
        "200":
          description: Matching messages; never persisted
          content:
            application/json:
              schema:
                type: object
                required: [messages, hasMore]
                properties:
                  hasMore: { type: boolean }
                  messages:
                    type: array
                    items:
                      type: object
                      required: [id, sender, subject, date, snippet]
                      properties:
                        id: { type: string }
                        sender: { type: string }
                        subject: { type: string }
                        date: { type: string }
                        snippet: { type: string }
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      responses:
        "200":
          description: All applications, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      operationId: createApplication
      tags: [applications]
      parameters:
        - in: header
          name: X-JobTrack-Gmail-Review
          required: false
          description: Signed review token from Gmail analysis; enables idempotent import in development only
          schema:
            type: string
            maxLength: 256
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "200":
          description: Matching application already exists; returned unchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /applications/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getApplication
      tags: [applications]
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      operationId: updateApplication
      tags: [applications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      operationId: deleteApplication
      tags: [applications]
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Application not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    GmailAnalysisInput:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: object
          additionalProperties: false
          required: [id, sender, subject, date, snippet]
          properties:
            id: { type: string, minLength: 1, maxLength: 256 }
            sender: { type: string, maxLength: 1000 }
            subject: { type: string, maxLength: 2000 }
            date: { type: string, maxLength: 100 }
            snippet: { type: string, maxLength: 10000 }
    GmailAnalysis:
      type: object
      required: [relevant, reason, extraction, reviewToken]
      properties:
        relevant: { type: boolean }
        reason: { type: string }
        reviewToken: { type: [string, "null"] }
        extraction:
          $ref: "#/components/schemas/GmailExtraction"
    GmailExtraction:
      type: object
      additionalProperties: false
      required: [companyName, position, opportunityType, status, applicationDate, interviewDate, interviewType, interviewLocation, interviewNotes, jobUrl]
      properties:
        companyName: { type: [string, "null"] }
        position: { type: [string, "null"] }
        opportunityType:
          type: [string, "null"]
          enum: [Full-time, Part-time, Internship, Training, Train-to-Hire, Freelance, null]
        status:
          type: [string, "null"]
          enum: [Saved, Applied, Under Review, Assessment, Interview, Technical Interview, Final Interview, Offer, Accepted, Rejected, Withdrawn, null]
        applicationDate: { type: [string, "null"] }
        interviewDate: { type: [string, "null"] }
        interviewType: { type: [string, "null"] }
        interviewLocation: { type: [string, "null"] }
        interviewNotes: { type: [string, "null"] }
        jobUrl: { type: [string, "null"] }
    GmailSearchInput:
      type: object
      properties:
        query:
          type: string
          maxLength: 500
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
    ApplicationInput:
      type: object
      additionalProperties: false
      required: [companyName, positionTitle, opportunityType, workMode, applicationDate, status, interviewDate, interviewTime, interviewType, interviewLocation, interviewNotes, jobUrl, notes, source]
      properties:
        companyName: { type: string, minLength: 1, maxLength: 255 }
        positionTitle: { type: string, minLength: 1, maxLength: 255 }
        opportunityType: { type: string, enum: [Full-time, Part-time, Internship, Training, Train-to-Hire, Freelance] }
        workMode: { type: string, enum: [On-site, Remote, Hybrid] }
        applicationDate: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$', description: Local calendar date (YYYY-MM-DD) }
        status: { type: string, enum: [Saved, Applied, Under Review, Assessment, Interview, Technical Interview, Final Interview, Offer, Accepted, Rejected, Withdrawn] }
        interviewDate: { type: string, description: Empty string or YYYY-MM-DD }
        interviewTime: { type: string, description: Empty string or HH:MM }
        interviewType: { type: string, maxLength: 255 }
        interviewLocation: { type: string, maxLength: 2048 }
        interviewNotes: { type: string }
        jobUrl: { type: string, maxLength: 2048, description: Empty string or an http(s) URL }
        notes: { type: string }
        source: { type: string, enum: [Manual, Gmail] }
    Application:
      type: object
      required: [id, companyName, positionTitle, opportunityType, workMode, applicationDate, status, interviewDate, interviewTime, interviewType, interviewLocation, interviewNotes, jobUrl, notes, source, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        companyName: { type: string, minLength: 1, maxLength: 255 }
        positionTitle: { type: string, minLength: 1, maxLength: 255 }
        opportunityType: { type: string, enum: [Full-time, Part-time, Internship, Training, Train-to-Hire, Freelance] }
        workMode: { type: string, enum: [On-site, Remote, Hybrid] }
        applicationDate: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        status: { type: string, enum: [Saved, Applied, Under Review, Assessment, Interview, Technical Interview, Final Interview, Offer, Accepted, Rejected, Withdrawn] }
        interviewDate: { type: string }
        interviewTime: { type: string }
        interviewType: { type: string, maxLength: 255 }
        interviewLocation: { type: string, maxLength: 2048 }
        interviewNotes: { type: string }
        jobUrl: { type: string, maxLength: 2048 }
        notes: { type: string }
        source: { type: string, enum: [Manual, Gmail] }
        createdAt: { type: string, description: Server-owned ISO 8601 timestamp }
        updatedAt: { type: string, description: Server-owned ISO 8601 timestamp }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status