# lib/api-zod/src/generated/types/gmailAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GmailExtraction } from './gmailExtraction';

export interface GmailAnalysis {
  relevant: boolean;
  reason: string;
  /** @nullable */
  reviewToken: string | null;
  extraction: GmailExtraction;
}