# lib/api-zod/src/generated/types/gmailAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GmailAnalysisInputMessage } from './gmailAnalysisInputMessage';

export interface GmailAnalysisInput {
  message: GmailAnalysisInputMessage;
}