# lib/api-zod/src/generated/types/gmailExtraction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GmailExtractionOpportunityType } from './gmailExtractionOpportunityType';
import type { GmailExtractionStatus } from './gmailExtractionStatus';

export interface GmailExtraction {
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  position: string | null;
  /** @nullable */
  opportunityType: GmailExtractionOpportunityType;
  /** @nullable */
  status: GmailExtractionStatus;
  /** @nullable */
  applicationDate: string | null;
  /** @nullable */
  interviewDate: string | null;
  /** @nullable */
  interviewType: string | null;
  /** @nullable */
  interviewLocation: string | null;
  /** @nullable */
  interviewNotes: string | null;
  /** @nullable */
  jobUrl: string | null;
}